requirements.txt
# ==============================================================================
# Avatar System Orchestrator - Complete Dependencies
# ==============================================================================

# Core Deep Learning
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# LangChain & Orchestration
langchain>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
langsmith>=0.1.0

# LLM APIs (Mistral & Groq)
mistralai>=0.1.0
groq>=0.4.0

# Transformers & Model Loading
transformers>=4.35.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Web Interfaces
gradio==4.13.0
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart>=0.0.6

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
openai-whisper>=20231117

# Video Processing (VidGear - Faster than MoviePy!)
vidgear[core]>=0.3.2
opencv-python==4.8.1.78
moviepy==1.0.3
imageio==2.31.5
imageio-ffmpeg==0.4.9
ffmpeg-python>=0.2.0

# Computer Vision & Face Processing
facenet-pytorch==2.5.3
mediapipe==0.10.11
kornia==0.7.0
insightface>=0.7.3
onnxruntime>=1.16.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
numpy==1.24.3
scipy==1.11.3
pillow==10.1.0
pandas==2.1.3
matplotlib==3.8.0
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0

# Colab & Deployment
gdown==4.7.1
huggingface-hub==0.19.4
google-colab==1.0.0
jupyter==1.0.0

# Monitoring & Logging
psutil==5.9.6
gputil==1.4.0
wandb==0.16.0
sentry-sdk>=1.38.0

# Observability & Tracing (Langfuse)
langfuse>=2.0.0
langfuse-langchain>=2.0.0

# Development & Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Optional (GPU Optimizations)
# Uncomment if you have compatible hardware:
# nvidia-cudnn-cu11>=8.9.0
# nvidia-cublas-cu11>=11.11.0
# triton>=2.1.0

# ==============================================================================
# Installation Notes:
# - Use: uv pip install -r requirements.txt
# - VidGear replaces Replicate API (local processing, faster!)
# - Mistral/Groq for LLM/Speech (API-based, no local download)
# - All models auto-download on first run
# ==============================================================================